chore: move modelUtil.ts into isomorphic/trace
