+++ packages/trace-viewer/src/ui/filmStrip.tsx
@@ -19,7 +19,7 @@ import type { Boundaries, Size } from './geometry';
 import * as React from 'react';
 import { useMeasure, upperBound } from '@web/uiUtils';
 import type { PageEntry } from '@isomorphic/trace/entries';
-import type { ActionTraceEventInContext } from './modelUtil';
+import type { ActionTraceEventInContext } from '@isomorphic/trace/traceModel';
 import { renderAction } from './actionList';
 import type { Language } from '@isomorphic/locatorGenerators';
 import { useTraceModel } from './traceModelContext';
